+++ extern/MOxUnit/MOxUnit/@MOxUnitTestSuite/countTestCases.m
@@ -12,8 +12,8 @@ function c=countTestCases(obj)
 % NNO 2015
 
     c=0;
-    for k=1:numel(obj.tests)
-        t=obj.tests{k};
+    for k=1:countTestNodes(obj)
+        t=getTestNode(obj,k);
         c=c+countTestCases(t);
     end